rename: 컴포넌트 디렉터리명 변경 (NavigateButton → NavigationButton)
