+++ src/pages/OrderConfirmPage.tsx
@@ -1,12 +1,23 @@
 import { Suspense } from "react";
+import { useNavigate } from "react-router-dom";
 import { ErrorBoundary } from "react-error-boundary";
 
-import { Header, Footer, OrderConfirm, Loading } from "@/components/index";
-import BackButton from "@/components/button/BackButton";
-import CartLayout from "@/layouts/Cart";
+import PurchaseLayout from "@/layouts/PurchaseLayout";
+
 import ErrorFallback from "@/components/_common/ErrorFallback";
+import { Header, Loading } from "@/components/index";
+import { NavigationButton, BackButton } from "@/components/button";
+import OrderContent from "@/components/order/OrderContent";
+
+import { ROUTE_PATH } from "@/constants/route";
 
 const OrderConfirmPage = () => {
+  const navigate = useNavigate();
+
+  const handleButtonClick = () => {
+    navigate(ROUTE_PATH.paymentConfirm);
+  };
+
   return (
     <>
       <Header>
@@ -14,10 +25,13 @@ const OrderConfirmPage = () => {
       </Header>
       <ErrorBoundary FallbackComponent={ErrorFallback}>
         <Suspense fallback={<Loading />}>
-          <CartLayout>
-            <OrderConfirm />
-          </CartLayout>
-          <Footer />
+          <PurchaseLayout title="주문 확인">
+            <OrderContent />
+          </PurchaseLayout>
+          <NavigationButton
+            buttonText="결제하기"
+            onButtonClick={handleButtonClick}
+          />
         </Suspense>
       </ErrorBoundary>
     </>
